requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.3.4
aiohttp==3.10.1
aiosignal==1.3.1
anaconda-anon-usage==0.4.4
annotated-types==0.7.0
anthropic==0.30.1
antlr4-python3-runtime==4.9.3
anyio==4.4.0
archspec==0.2.3
attrs==24.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
boltons==23.0.0
boto3==1.34.155
botocore==1.34.155
Brotli==1.0.9
cachetools==5.3.3
certifi==2024.7.4
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
circuitbreaker==1.4.0
click==8.1.7
cn2an==0.5.22
cohere==5.5.8
coloredlogs==15.0.1
conda==24.5.0
conda-content-trust==0.2.0
conda-libmamba-solver==24.1.0
conda-package-handling==2.2.0
conda_package_streaming==0.9.0
contourpy==1.2.1
cryptography==42.0.8
cssselect==1.2.0
cycler==0.12.1
dataclasses-json==0.6.7
datrie==0.8.2
deepdiff==7.0.1
defusedxml==0.7.1
Deprecated==1.2.14
distro==1.9.0
effdet==0.4.1
elastic-transport==8.13.1
elasticsearch==8.12.1
emoji==2.12.1
et-xmlfile==1.1.0
fastapi==0.112.0
fastavro==1.9.5
ffmpy==0.4.0
filelock==3.15.4
filetype==1.2.0
flatbuffers==24.3.25
fonttools==4.53.1
frozendict==2.4.2
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.1
google-auth==2.33.0
google-cloud-vision==3.7.4
googleapis-common-protos==1.63.2
gradio==4.39.0
gradio_client==1.1.1
greenlet==3.0.3
grpcio==1.65.4
grpcio-status==1.65.4
h11==0.14.0
hanziconv==0.3.2
html_text==0.6.2
httpcore==1.0.5
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.24.5
humanfriendly==10.0
idna==3.7
importlib_resources==6.4.0
iopath==0.1.10
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
kiwisolver==1.4.5
langchain==0.2.6
langchain-anthropic==0.1.19
langchain-cohere==0.1.8
langchain-community==0.2.6
langchain-core==0.2.28
langchain-openai==0.1.11
langchain-text-splitters==0.2.2
langdetect==1.0.9
langfuse==2.53.9
langsmith==0.1.98
layoutparser==0.3.4
libmambapy==1.5.8
llvmlite==0.42.0
lxml==5.2.2
lxml_html_clean==0.2.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.1.post1
mdurl==0.1.2
menuinst==2.0.2
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numba==0.59.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
oci==2.129.0
olefile==0.47
omegaconf==2.3.0
onnx==1.16.2
onnxruntime==1.18.1
onnxruntime-gpu==1.18.1
openai==1.40.0
opencv-python==4.9.0.80
openpyxl==3.1.2
oracledb==2.2.1
ordered-set==4.1.0
orjson==3.10.6
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.2
parameterized==0.9.0
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.2
pikepdf==9.1.0
pillow==10.3.0
pillow_heif==0.18.0
pip==24.0
platformdirs==3.10.0
pluggy==1.0.0
portalocker==2.10.1
proces==0.1.7
proto-plus==1.24.0
protobuf==5.27.3
psutil==6.0.0
py-cpuinfo==9.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyclipper==1.3.0.post5
pycocotools==2.0.8
pycosat==0.6.6
pycparser==2.22
pycryptodome==3.20.0
pycryptodomex==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
Pygments==2.18.0
PyMuPDF==1.24.7
PyMuPDFb==1.24.6
pyOpenSSL==24.2.1
pypandoc==1.13
pyparsing==3.1.2
pypdf==4.3.1
PyPDF2==3.0.1
pypdfium2==4.30.0
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2024.4.27
python-magic==0.4.27
python-multipart==0.0.9
python-oxmsg==0.0.1
python-pptx==0.6.23
pytz==2024.1
PyYAML==6.0.2
rapidfuzz==3.9.6
readability-lxml==0.8.1
regex==2024.7.24
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.7.1
roman-numbers==1.0.2
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
ruff==0.5.6
s3transfer==0.10.2
safetensors==0.4.4
scipy==1.14.0
seaborn==0.13.2
selenium==4.21.0
semantic-version==2.10.0
setuptools==72.1.0
shapely==2.0.3
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.32
starlette==0.37.2
StrEnum==0.4.15
sympy==1.13.1
tabulate==0.9.0
tenacity==8.5.0
thop==0.1.1.post2209072238
tiktoken==0.7.0
timm==1.0.8
tokenizers==0.19.1
tomlkit==0.12.0
torch==2.2.1+cu121
torchaudio==2.2.1+cu121
torchvision==0.17.1+cu121
tqdm==4.66.4
transformers==4.44.0
trio==0.26.1
trio-websocket==0.11.1
triton==2.2.0
truststore==0.8.0
typer==0.12.3
types-requests==2.32.0.20240712
typing_extensions==4.12.2
typing-inspect==0.9.0
tzdata==2024.1
ultralytics==8.2.93
ultralytics-thop==2.0.6
unstructured==0.14.10
unstructured-client==0.25.4
unstructured-inference==0.7.36
unstructured.pytesseract==0.3.12
urllib3==2.2.2
uvicorn==0.30.5
webdriver-manager==4.0.1
websockets==11.0.3
Werkzeug==3.0.3
wheel==0.43.0
word2number==1.1
wrapt==1.16.0
wsproto==1.2.0
xgboost==2.1.0
xlrd==2.0.1
XlsxWriter==3.2.0
yarl==1.9.4
zstandard==0.22.0
